New file to deal with hand tracking and landmarker
Added the new landmarker file for hands, and created a new file that's
purpose is to deal with the hand annotation stuff. I guess the coloring
stuff will have its own file, maybe I'll move all webcam related stuff
there.

The next step is to detect when which fingers are up. This might be done
in the hand function?