Started implementing the basics of Selection Mode
The program can now detect when a hand is in the selection area (for y,
x still needs to be implemented). It has a for loop it goes through
every frame though, which I don't really like... I want to find a more
efficient way to do this.

Next step is to determine specifically which color box it's in by
looking at the x values.