Drawing now works
The squares that are drawn are a little too dark, though.
This is because they are drawn on a new overlay, and then
the banner is put on that overlay, and when that happens the alpha
levels are changed for the first overlay drawing, causing the squares to
have a low alpha value. When I tried to turn it up, the image got
distorted, but that might have been because of the light level in my
room? So maybe I'll try this at another time when the sun isn't beaming
so brightly and see if it looks better.