"Fixed" issue 12?
Changed the input to the function that recognizes the hand and gave it
the frame_as_img (which converts color) rather than just the frame that
is returned from cv2. It seems to be a lot quicker with recognizing when
the hand comes back on screen now. Now sure if this is necessarily
fixed or if it's just working a little better...