Blending works & added cleaner way to add colors
Blending works decently now. If radius is too high, it can be hard to
blend, because the color gets placed too far away from the finger before
the user is able to touch it to blend.

Also, the colors are now loaded from a loop instead of individually
having to call the load and process function for each one.

Also also I added green, white, and black for some more variety in the
colors. White and black so users can make colors brighter or darker,
green since pixels are RGB so this might be a better way to blend than
using primary colors like in actual art.